+++ CHANGELOG.md
@@ -10,6 +10,9 @@ All notable changes to this project will be documented in this file.
 - Added Safehouses dumper.
 - Added protection from changes to logutils.
 
+## Changed
+- Changed format of character's coordinate in brushtool, craft, player and safehouse logs. Old:`.. (x,y,z).`, new `.. at x,y,z.` 
+
 ## [v0.14.0] - 2026-08-31
 ### Added
 - Added skipping player tick log when character is dead.